mesas alter table en migracion

+++ database/migrations/2023_04_09_091737_create_productos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nombre');
             $table->integer('existencias');
             $table->string('alergenos');
-            $table->integer('precio');
+            $table->double('precio');
             $table->string('descripcion');
             $table->string('ingredientes');
             $table->string('imagen');
